move doctomarkdown to indexer-internal file
